fix: satisfy clippy for v1.1.0 release workflow
Collapse nested ifs, simplify float finiteness checks, and drop
unnecessary casts so release CI passes -D warnings.

+++ crates/propaga-propagators/src/int_mod.rs
@@ -105,7 +105,7 @@ impl Propagator for IntModPropagator {
             changed |= ctx.remove_below(c, 0);
             changed |= ctx.remove_above(c, bmin - 1);
         } else if bmax < 0 {
-            let abs_lo = (-bmax) as i32;
+            let abs_lo = -bmax;
             changed |= ctx.remove_above(c, 0);
             changed |= ctx.remove_below(c, 1 - abs_lo);
         } else if bmin < 0 && bmax > 0 {
